+++ NEWS.md
@@ -1,6 +1,8 @@
 # rmd4j 0.0.0.9000
 
-* moved set_neo4J_engine to set_neo4j_engine (no uppercase)
+* 2018-03-06 : you now have to give your NEO4JHOME, not the path to your bin.
+
+* 2018-03-02 : changed set_neo4J_engine to set_neo4j_engine (no uppercase)
 
 * First draft of the package